Add interact.js
