+++ src/index.html
@@ -8,6 +8,15 @@
   <script src="https://kit.fontawesome.com/e6861a11cf.js" crossorigin="anonymous"></script>
 </head>
 <body>
-  <div id="todolist"></div>
+  <div class="todolist">
+    <div>
+      <h1>Today's To dos</h1>
+    </div>
+    <div>
+      <input id="newTask" type="text">
+    </div>
+    <div id="todolist"></div>
+    <div></div>
+  </div>
 </body>
 </html>
